+++ tests/fixtures/unsupported_object_features/unsupported_clone_expression.php
@@ -1,0 +1,6 @@
+<?php
+class Box {
+    public $name;
+}
+$box = new Box();
+$copy = clone $box;